combine ProjectTableData and SubmissionTableData components, move to UI folder
