+++ client/src/components/Submissions/SubmissionTableRow.jsx
@@ -15,7 +15,7 @@ import AdminNotesModal from "../Modals/ActionProjectAdminNotes";
 import ActionManageSubmission from "../Modals/ActionManageSubmission";
 import WarningModal from "../Modals/WarningAdminNotesUnsavedChanges";
 import InfoSubmissionLog from "../Modals/InfoSubmissionLog";
-import { Td, TdExpandable } from "./SubmissionTableData";
+import { Td, TdExpandable } from "../UI/TableData";
 
 function useAdminNotesModal(project, onAdminNoteUpdate) {
   const [showWarningModal, setShowWarningModal] = useState(false);
